Add tedge_utils::fs atomically write file
