+++ common/tedge_utils/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod fs;
 pub mod paths;
 pub mod signals;
