+++ tedge_config/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 certificate = { path = "../common/certificate" }
 serde = { version = "1.0", features = ["derive"] }
+tedge_utils = { path = "../common/tedge_utils" }
 tempfile = "3.2"
 thiserror = "1.0"
 toml = "0.5"
